break out POST_LiccenseSerializer
