+++ api/serializers/__init__.py
@@ -22,3 +22,4 @@ from .paginated_instance_history_serializer import PaginatedInstanceHistorySeria
 from .paginated_instance_serializer import PaginatedInstanceSerializer
 from .machine_export_serializer import MachineExportSerializer
 from .license_serializer import LicenseSerializer
+from .post_license_serializer import POST_LicenseSerializer
